端到端实测暴露的修复：复权一致性/ROA口径/样本警告/float强转
apply_adj_factor：改为一并复权 OHLC(open/high/low/close/pre_close)，*_post 同 scale，
  避免只复权 close 导致 KDJ 等 OHLC 类指标 scale 不一致(实测 002500 算出 K=735)
SKILL.md：daily 实为未复权(原文误标前复权)，趋势/技术指标必须复权；补 index_weight
  月度+整月范围规范、stock_company/cn_cpi(nt_yoy)/cn_ppi(ppi_yoy)/margin(exchange_id) 字段口径
F-Score：ROA 取 npta 优先(金融业 roa 字段为空)、次 roa、末用净利润/总资产自算；
  comp_type 2/3/4 触发金融业弱参考 comp_note；新增 有效项/缺失项 透明计数
valuation_percentiles：新增 截面样本数/截面可靠性，<10 标注样本不足
calc_*：dict 数值统一 float() 强转，消除 np.float64 泄露
测试：新增 test_fscore_financial_firm_comp_note(npta 兜底+comp_note)、扩展 valuation 样本断言

+++ sxsc-tushare-analysis/scripts/basic_metrics.py
@@ -15,10 +15,10 @@ def calc_returns(df_close, periods=(5, 20, 60, 120, 250)):
     不要用整数 p 直接访问（如 out["近20日涨幅%"]，而非 out[20]）。
     """
     latest = df_close.iloc[-1]
-    out = {"最新收盘": round(latest, 2)}
+    out = {"最新收盘": round(float(latest), 2)}
     for p in periods:
         if len(df_close) > p:
-            out[f"近{p}日涨幅%"] = round((latest / df_close.iloc[-1 - p] - 1) * 100, 2)
+            out[f"近{p}日涨幅%"] = round(float((latest / df_close.iloc[-1 - p] - 1) * 100), 2)
     return out
 
 
@@ -35,7 +35,7 @@ def calc_cagr(start_value, end_value, days):
 
 def calc_ma(df_close, windows=(5, 10, 20, 60)):
     """均线值（用于判断多头/空头排列）。"""
-    return {f"MA{w}": round(df_close.tail(w).mean(), 2) for w in windows}
+    return {f"MA{w}": round(float(df_close.tail(w).mean()), 2) for w in windows}
 
 
 def calc_volatility(df_close, window=20):
